Add support for int output to `.ensime` files
Certain items in :formatting-prefs are supposed to be integers, not
strings, so add support for these.

Specific item that was causing issues for me was `indentSpaces` in this
project itself, which would print as `:indentSpaces "2"`, which unless I
deleted that particular line would prevent ensime from starting up to
hack around on this project. With this change, it now goes into the
`.ensime` file as `:indentSpaces 2`.