Configure scalariform to run automatically
In my last PR (https://github.com/ensime/ensime-maven/pull/13), I
mentioned that we should actually run scalariform on this project. So,
here that is. The config is taken from
https://github.com/tanzoniteblack/ensime-maven/blob/scalariform/src/ensime/formatter.properties
, but I actually have no strong preferences one way or the other on any
of these values.

Note: The way that I set up the pom.xml, it will run scalariform
automatically when you compile the code.