Ignore pom artifact types
One of my projects has a dependency that looks like:

```xml
<dependency>
	<groupId>com.github.fommil.netlib</groupId>
	<artifactId>all</artifactId>
	<version>1.1.2</version>
	<type>pom</type>
</dependency>
```

This ends up going into the `.ensime` file along with all the dependency
jars, and then causes ensime to throw an error because it's not a zip
file. So let's just avoid putting anything into the the source roots
unless it's a jar type file according to the artifact data.